+++ src/server/Routes.ts
@@ -9,6 +9,7 @@ import * as PlaceRoute from './Place/PlaceRoute';
 import * as PresetRoute from './Preset/PresetRoute';
 import * as TaskRoute from './Task/TaskRoute';
 import * as LastTimeRoute from './LastTime/LastTimeRoute';
+import * as SyncRoute from './Sync/SyncRoute';
 
 // Internal
 import * as TypeGenerator from './TypeGenerator/TypeGeneratorRoute';
@@ -25,6 +26,8 @@ router.use('/places', PlaceRoute.router);
 router.use('/presets', PresetRoute.router);
 router.use('/tasks', TaskRoute.router);
 router.use('/lasttime', LastTimeRoute.router);
+
+router.use('/sync', SyncRoute.router);
 
 // Routing for internals
 router.use('/type-generator', TypeGenerator.router);